--- frontend/src/components/AddMovie/AddMovie.js
@@ -1,14 +1,0 @@
-import React from "react";
-import { AddMovieButtonContainer, AddMovieButton } from "./AddMovieStyles";
-
-const AddMovie = ({ addMovieModal }) => {
-  return (
-    <AddMovieButtonContainer>
-      <AddMovieButton red onClick={addMovieModal}>
-        + ADD MOVIE
-      </AddMovieButton>
-    </AddMovieButtonContainer>
-  );
-};
-
-export default AddMovie;
