+++ frontend/src/components/App/App.js
@@ -1,60 +1,57 @@
-import React, { useState } from "react";
+import React, { useState, useCallback } from "react";
 import { GlobalStyle } from "./GlobalStyles";
-import MovieData from "../../mocks/Data";
 import Home from "../../pages/Home";
+import ErrorBoundary from "./ErrorBoundary";
+import { Provider } from "react-redux";
+import { BrowserRouter } from "react-router-dom";
 
-export default function App() {
-  const [movieData, setMovieData] = useState(MovieData);
+export default function App({ store }) {
   const [movieModal, setMovieModal] = useState(false);
   const [modalTitle, setModalTitle] = useState("");
   const [deleteModal, setDeleteModal] = useState(false);
   const [switchBanner, setSwitchBanner] = useState(false);
   const [movieDetailsId, setMovieDetailsId] = useState();
 
-  function closeModal(e) {
-    // console.log(e.target.tagName);
-    if (e.target.tagName === "DIV") {
-      setMovieModal(false);
-      // setDeleteModal(false);
-    }
-  }
+  const closeModal = useCallback(
+    (e) => {
+      if (e.target.tagName === "DIV") {
+        setMovieModal(false);
+      }
+    },
+    [movieModal]
+  );
 
-  function addMovieModal() {
+  const addMovieModal = useCallback(() => {
     setMovieModal(true);
     setModalTitle("ADD MOVIE");
-  }
+  }, [movieModal]);
 
-  function editMovieModal() {
+  const editMovieModal = useCallback(() => {
     setMovieModal(true);
     setModalTitle("EDIT MOVIE");
-  }
-
-  function handleMovieDelete(id) {
-    if (movieDetailsId !== null && movieDetailsId === id) {
-      setMovieDetailsId(undefined);
-    }
-    setMovieData(movieData.filter((recipe) => recipe.id !== id));
-  }
+  }, [movieModal, modalTitle]);
 
   return (
-    <div>
-      <GlobalStyle />
-      <Home
-        handleMovieDelete={handleMovieDelete}
-        movieData={movieData}
-        editMovieModal={editMovieModal}
-        setMovieModal={setMovieModal}
-        modalTitle={modalTitle}
-        movieDetailsId={movieDetailsId}
-        setMovieDetailsId={setMovieDetailsId}
-        movieModal={movieModal}
-        addMovieModal={addMovieModal}
-        deleteModal={deleteModal}
-        setDeleteModal={setDeleteModal}
-        closeModal={closeModal}
-        switchBanner={switchBanner}
-        setSwitchBanner={setSwitchBanner}
-      />
-    </div>
+    <Provider store={store}>
+      <BrowserRouter>
+        <ErrorBoundary>
+          <GlobalStyle />
+          <Home
+            editMovieModal={editMovieModal}
+            setMovieModal={setMovieModal}
+            modalTitle={modalTitle}
+            movieDetailsId={movieDetailsId}
+            setMovieDetailsId={setMovieDetailsId}
+            movieModal={movieModal}
+            addMovieModal={addMovieModal}
+            deleteModal={deleteModal}
+            setDeleteModal={setDeleteModal}
+            closeModal={closeModal}
+            switchBanner={switchBanner}
+            setSwitchBanner={setSwitchBanner}
+          />
+        </ErrorBoundary>
+      </BrowserRouter>
+    </Provider>
   );
 }
