--- frontend/src/components/Movies/MoviesList/MoviesList.js
@@ -1,20 +1,0 @@
-import React, { useState } from "react";
-import Movies from "../Movies/Movies";
-import FilterMovies from "../../FilterMovies/FilterMovies";
-import { MoviesListContainer, Container } from "./MoviesListStyles";
-
-const MoviesList = ({ setSwitchBanner, editMovieModal }) => {
-  return (
-    <MoviesListContainer>
-      <Container>
-        <FilterMovies />
-        <Movies
-          editMovieModal={editMovieModal}
-          setSwitchBanner={setSwitchBanner}
-        />
-      </Container>
-    </MoviesListContainer>
-  );
-};
-
-export default MoviesList;
