+++ frontend/src/components/shared/Footer/index.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Icon from "../Icon/Icon";
+import Icon from "../Icon";
 import { FooterContainer } from "./FooterStyles";
 
 const Footer = () => {
