+++ frontend/src/index.js
@@ -1,15 +1,6 @@
 import React from "react";
 import { render } from "react-dom";
 import App from "./components/App/App";
-import { Provider } from "react-redux";
-import { BrowserRouter } from "react-router-dom";
 import store from "./store/store";
 
-render(
-  <Provider store={store}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </Provider>,
-  document.getElementById("app")
-);
+render(<App store={store} />, document.getElementById("app"));
