--- frontend/src/mocks/Data.js
@@ -1,99 +1,0 @@
-const movies = [
-  {
-    id: 1,
-    title: "Pulp Fiction",
-    description:
-      "Pulp Fiction is a 1994 American neo-noir black comedy crime film written and directed by Quentin Tarantino, who conceived it with Roger Avary. Starring John Travolta, Samuel L. Jackson, Bruce Willis, Tim Roth, Ving Rhames, and Uma Thurman, it tells several stories of criminal Los Angeles. The title refers to the pulp magazines and hardboiled crime novels popular during the mid-20th century, known for their graphic violence and punchy dialogue.",
-    image:
-      "//upload.wikimedia.org/wikipedia/en/3/3b/Pulp_Fiction_%281994%29_poster.jpg",
-    genre: "Crime",
-    year: 2004,
-    rating: 8.9,
-    runtime: 178,
-  },
-  {
-    id: 2,
-    title: "Bohemian Rhapsody",
-    description:
-      "Bohemian Rhapsody is a 2018 biographical drama film about Freddie Mercury, the lead singer of the British rock band Queen. It was directed by Bryan Singer[a] from a screenplay by Anthony McCarten, and produced by Graham King and Queen manager Jim Beach. It stars Rami Malek as Mercury, with Lucy Boynton, Gwilym Lee, Ben Hardy, Joe Mazzello, Aidan Gillen, Tom Hollander, Allen Leech, and Mike Myers in supporting roles. Queen members Brian May and Roger Taylor served as consultants. A British-American venture, the film was produced by 20th Century Fox, Regency Enterprises, GK Films, and Queen Films, with Fox serving as distributor. The film follows the singer's life from the formation of the band up to their 1985 Live Aid performance at the original Wembley Stadium.",
-    image:
-      "https://upload.wikimedia.org/wikipedia/en/thumb/2/2e/Bohemian_Rhapsody_poster.png/220px-Bohemian_Rhapsody_poster.png",
-    genre: "Classic Rock",
-    year: 2018,
-    rating: 8.0,
-    runtime: 132,
-  },
-  {
-    id: 3,
-    title: "Kill Bill",
-    description:
-      "Kill Bill: Volume 2 is a 2004 American martial arts film written and directed by Quentin Tarantino. It stars Uma Thurman as the Bride, who continues her campaign of revenge against the Deadly Viper Assassination Squad (Lucy Liu, Michael Madsen, Daryl Hannah, and Vivica A. Fox) and their leader Bill (David Carradine), who tried to kill her and her unborn child.",
-    image:
-      "https://upload.wikimedia.org/wikipedia/en/c/c4/Kill_Bill_Volume_2.png",
-    genre: "Action",
-    year: 1994,
-    rating: 8.1,
-    runtime: 112,
-  },
-  {
-    id: 4,
-    title: "Avengers: Infinity War",
-    description:
-      "Avengers: Infinity War is a 2018 American superhero film based on the Marvel Comics superhero team the Avengers, produced by Marvel Studios and distributed by Walt Disney Studios Motion Pictures. It is the sequel to The Avengers (2012) and Avengers: Age of Ultron (2015), and the 19th film in the Marvel Cinematic Universe (MCU).",
-    image:
-      "https://upload.wikimedia.org/wikipedia/en/4/4d/Avengers_Infinity_War_poster.jpg",
-    genre: "Science Fiction",
-    year: 2018,
-    rating: 8.4,
-    runtime: 160,
-  },
-  {
-    id: 5,
-    title: "Inception",
-    description:
-      "Inception is a 2010 science fiction action film[4][5][6] written and directed by Christopher Nolan, who also produced the film with his wife, Emma Thomas. The film stars Leonardo DiCaprio as a professional thief who steals information by infiltrating the subconscious of his targets. He is offered a chance to have his criminal history erased as payment for the implantation of another person's idea into a target's subconscious.[7] The ensemble cast includes Ken Watanabe, Joseph Gordon-Levitt, Marion Cotillard, Elliot Page,[a] Tom Hardy, Dileep Rao, Cillian Murphy, Tom Berenger, and Michael Caine.",
-    image:
-      "https://upload.wikimedia.org/wikipedia/en/2/2e/Inception_%282010%29_theatrical_poster.jpg",
-    genre: "Thriller",
-    year: 2003,
-    rating: 8.8,
-    runtime: 162,
-  },
-  {
-    id: 6,
-    title: "Reservoir Dogs",
-    description:
-      "Reservoir Dogs is a 1992 American crime film written and directed by Quentin Tarantino in his feature-length debut. It stars Harvey Keitel, Tim Roth, Chris Penn, Steve Buscemi, Lawrence Tierney, Michael Madsen, Quentin Tarantino, and Edward Bunker as diamond thieves whose planned heist of a jewelry store goes terribly wrong. The film depicts the events before and after the heist. Kirk Baltz, Randy Brooks, and Steven Wright also play supporting roles. It incorporates many motifs that have become Tarantino's hallmarks: violent crime, pop culture references, profanity, and nonlinear storytelling.",
-    image: "https://upload.wikimedia.org/wikipedia/en/0/01/Reservoir_Dogs.png",
-    genre: "Crime",
-    year: 1994,
-    rating: 8.3,
-    runtime: 100,
-  },
-  {
-    id: 7,
-    title: "The Hurt Locker",
-    description:
-      "The Hurt Locker is a 2008 American war thriller film directed by Kathryn Bigelow and written by Mark Boal. It stars Jeremy Renner, Anthony Mackie, Brian Geraghty, Christian Camargo, Ralph Fiennes, David Morse, and Guy Pearce. The film follows an Iraq War Explosive Ordnance Disposal team who are targeted by insurgents, and shows their psychological reactions to the stress of combat, which is intolerable to some and addictive to others. Boal drew on his experience during embedded access to write the screenplay.",
-    image:
-      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcR3wTlYKzTCYSxBomq6MQxAkWEORV3ugdTNmOY6gn-q6EosVQw8",
-    genre: "Action",
-    year: 2008,
-    rating: 7.5,
-    runtime: 133,
-  },
-  {
-    id: 8,
-    title: "Ready Player One",
-    description:
-      "Ready Player One is a 2018 American science fiction action-adventure film directed by Steven Spielberg, from a screenplay by Zak Penn and Ernest Cline, based on Cline's 2011 novel of the same name. It stars Tye Sheridan, Olivia Cooke, Ben Mendelsohn, Lena Waithe, T.J. Miller, Simon Pegg, and Mark Rylance, and the film takes place in 2045 when much of humanity uses the virtual reality software to escape the real world. Orphaned teenager Wade Watts (Sheridan) finds clues to a hidden game that promises OASIS's ownership to the winner, and he and four allies try to complete it before a corporation run by businessman Nolan Sorrento (Mendelsohn) can do so.",
-    image:
-      "https://upload.wikimedia.org/wikipedia/en/7/74/Ready_Player_One_%28film%29.png",
-    genre: "Science Fiction",
-    year: 2018,
-    rating: 7.5,
-    runtime: 139,
-  },
-];
-
-export default movies;
