+++ frontend/src/pages/Home.js
@@ -1,11 +1,9 @@
 import React from "react";
-import Banner from "../components/shared/Banner/Banner";
-import Footer from "../components/shared/Footer/Footer";
-import MoviesList from "../components/Movies/MoviesList/MoviesList";
+import Banner from "../components/shared/Banner";
+import Footer from "../components/shared/Footer";
+import MoviesList from "../components/Movies/MoviesList";
 
 const Home = ({
-  deleteModal,
-  setDeleteModal,
   movieModal,
   addMovieModal = { addMovieModal },
   closeModal,
@@ -13,11 +11,8 @@ const Home = ({
   switchBanner,
   setSwitchBanner,
   movieDetailsId,
-  setMovieDetailsId,
   modalTitle,
   editMovieModal,
-  movieData,
-  handleMovieDelete,
 }) => {
   return (
     <div>
@@ -32,14 +27,8 @@ const Home = ({
         closeModal={closeModal}
       />
       <MoviesList
-        handleMovieDelete={handleMovieDelete}
-        movieData={movieData}
         editMovieModal={editMovieModal}
-        setMovieDetailsId={setMovieDetailsId}
         setSwitchBanner={setSwitchBanner}
-        deleteModal={deleteModal}
-        setDeleteModal={setDeleteModal}
-        closeModal={closeModal}
       />
       <Footer />
     </div>
