+++ frontend/src/store/actions/constants.js
@@ -1,0 +1,14 @@
+export const ADD_MOVIE = "ADD_MOVIE";
+export const DELETE_MOVIE = "DELETE_MOVIE";
+export const EDIT_MOVIE = "EDIT_MOVIE";
+export const INIT_MOVIE_DETAILS = "INIT_MOVIE_DETAILS";
+export const FETCH_MOVIES = "FETCH_MOVIES";
+export const GET_MOVIE_DETAILS = "GET_MOVIE_DETAILS";
+export const FETCH_SEARCHED_MOVIES = "FETCH_SEARCHED_MOVIES";
+export const FETCH_SORTED_MOVIES = "FETCH_SORTED_MOVIES";
+export const SORT_BY_TITLE = "SORT_BY_TITLE";
+export const SORT_BY_RATING = "SORT_BY_RATING";
+export const SORT_BY_RELEASE_DATE = "SORT_BY_RELEASE_DATE";
+export const SHOW_LOADER = "SHOW_LOADER";
+export const HIDE_LOADER = "HIDE_LOADER";
+export const INIT_FETCH_MOVIES = "INIT_FETCH_MOVIES";
