+++ package.json
@@ -12,20 +12,31 @@
     "client": "npm start --prefix frontend",
     "dev": "concurrently \"npm run server\" \"npm run client\"",
     "watch": "nodemon --exec npm run start",
-    "lint": "eslint ./src",
+    "lint": "eslint .",
+    "lint:fix": "eslint .",
     "test": "jest"
   },
   "devDependencies": {
     "@babel/preset-env": "^7.8.4",
+    "babel-eslint": "^10.1.0",
     "concurrently": "^5.3.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.8.1",
+    "eslint-config-airbnb": "^18.2.0",
+    "eslint-config-prettier": "^6.11.0",
     "eslint-config-standard": "^14.1.0",
-    "eslint-plugin-import": "^2.20.0",
+    "eslint-config-wesbos": "^1.0.1",
+    "eslint-plugin-html": "^6.1.0",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-node": "^11.0.0",
+    "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-react": "^7.20.6",
+    "eslint-plugin-react-hooks": "^4.1.2",
     "eslint-plugin-standard": "^4.0.1",
     "jest": "^25.1.0",
     "nodemon": "^2.0.6",
+    "prettier": "^2.1.1",
     "webpack": "^5.12.2",
     "webpack-cli": "^4.3.1"
   },
